fix(power-calculation): Solve issues related to calculation of the produced power after sorting the powerplants by cost per MWh.

+++ src/PowerplantCodingChallenge.WebApi/Models/CalculatePowerProductionPlanRequest.cs
@@ -5,7 +5,7 @@ namespace PowerplantCodingChallenge.WebApi.Models;
 public record CalculatePowerProductionPlanRequest
 {
   [JsonPropertyName("load")]
-  public decimal LoadInMWh { get; init; }
+  public double LoadInMWh { get; init; }
 
   public Fuels Fuels { get; init; } = null!;
 
@@ -16,31 +16,31 @@ public record CalculatePowerProductionPlanRequest
 public record Fuels
 {
   [JsonPropertyName("gas(euro/MWh)")]
-  public float GasPriceInEuroPerMWh { get; init; }
+  public double GasPriceInEuroPerMWh { get; init; }
 
 
   [JsonPropertyName("kerosine(euro/MWh)")]
-  public float KerosinePriceInEuroPerMWh { get; init; }
+  public double KerosinePriceInEuroPerMWh { get; init; }
 
 
   [JsonPropertyName("co2(euro/ton)")]
-  public float Co2EmissionAllowancesPriceInEuroPerTon { get; init; }
+  public double Co2EmissionAllowancesPriceInEuroPerTon { get; init; }
 
 
   [JsonPropertyName("wind(%)")]
-  public float WindPercentage { get; init; }
+  public double WindPercentage { get; init; }
 }
 
 public record PowerPlant
 {
   public string Name { get; init; } = null!;
   public string Type { get; init; } = null!;
-  public float Efficiency { get; init; }
+  public double Efficiency { get; init; }
 
   [JsonPropertyName("pmin")]
-  public uint MinimumPowerOutput { get; init; }
+  public double MinimumPowerOutput { get; init; }
 
   [JsonPropertyName("pmax")]
-  public uint MaximumPowerOutput { get; init; }
+  public double MaximumPowerOutput { get; init; }
 }
