+++ src/PowerplantCodingChallenge.WebApi/Models/CalculatePowerProductionPlanResponse.cs
@@ -8,5 +8,5 @@ public record CalculatePowerProductionPlanResponse
   public string PowerPlantName { get; init; } = null!;
 
   [JsonPropertyName("p")]
-  public decimal PowerProducedInMWh { get; init; }
+  public double PowerProducedInMWh { get; init; }
 }
